% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bwght}
\alias{bwght}
\title{bwght}
\format{
A data frame with 41 observations.
\describe{
\item{faminc}{1988 family income, $1000s}
\item{cigtax}{cig. tax in home state, 1988}
\item{cigprice}{cig. price in home state, 1988}
\item{bwght}{birth weight, ounces}
\item{fatheduc}{father's yrs of educ}
\item{motheduc}{mother's yrs of educ}
\item{parity}{birth order of child}
\item{male}{=1 if male child}
\item{white}{=1 if white}
\item{cigs}{cigs smked per day while preg}
}
}
\source{
Wooldrige (2006): \emph{Introductory Econometrics: A Modern Approach},
3rd ed., Thomson South-Western.
}
\usage{
data("bwght")
}
\description{
Birth weight and cigarette smoking.
}
\details{
Data from J. Mullahy (1997), “Instrumental-Variable Estimation of
Count Data Models: Applications to Models of Cigarette Smoking Behavior,”
\emph{Review of Economics and Statistics} 79, 596-593. Professor Mullahy kindly
provided the data. He obtained them from the 1988 National Health
Interview Survey.
}
\keyword{datasets}
