% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop_test.R
\name{drop_test}
\alias{drop_test}
\title{Wald test for the exclusion of regressors}
\usage{
drop_test(model, frml = NULL, vce = NULL, chisq = FALSE)
}
\arguments{
\item{model}{an estimated model returned by \code{lm} or similar functions.}

\item{frml}{an one-sided formula specifying the terms to be tested.
If \code{NULL} all of the model regressors are used.}

\item{vce}{an object indicating how to obtain the covariance matrix.}

\item{chisq}{if \code{TRUE} the \eqn{\chi^2} version is returned.
By default is \code{FALSE} and the F version is computed.}
}
\value{
An object of class \code{htest} with components:
\describe{
\item{statistic}{the value of the test statistic.}
\item{p.value}{the p-value of the test.}
\item{parameter}{degrees of freedom.}
\item{method}{a character string indicating what type of test was
performed.}
\item{data.name}{a character string describing the model.}
}
}
\description{
Test that variables in \code{frml} can be excluded from \code{model} using a Wald test.
The parameter \code{vce} determines how is computed the covariance matrix of the
model parameters (see \link[=se]{se()}).
}
\examples{
data("hprice1")

mod <- lm(price ~ sqrft + bdrms + colonial, data = hprice1)

# Heteroskedasticity-robust joint significance test
drop_test(mod, vce = "HC")

# Test that bdrms and colonial are redundant (heteroskedasticity robust)
drop_test(mod, ~ bdrms + colonial, vce = "HC")

}
