% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{earns}
\alias{earns}
\title{earns}
\format{
A data frame with 41 observations.
\describe{
\item{year}{1947 to 1987}
\item{wkearns}{avg. real weekly earnings}
\item{wkhours}{avg. weekly hours}
\item{outphr}{output per labor hour}
}
}
\source{
Wooldrige (2006): \emph{Introductory Econometrics: A Modern Approach},
3rd ed., Thomson South-Western.
}
\usage{
data("earns")
}
\description{
Earnings, productivity and hours (macro).
}
\details{
Collected from the \emph{Economic Report of the President}, 1989, Table B-47.
The data are for the non-farm business sector.
}
\section{Notes}{

These data could be usefully updated, but changes in reporting conventions
in more recent ERPs may make that difficult.
}

\keyword{datasets}
