% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gpa1}
\alias{gpa1}
\title{gpa1}
\format{
A data frame with 141 observations.
\describe{
\item{age}{in years}
\item{soph}{=1 if sophomore}
\item{junior}{=1 if junior}
\item{senior}{=1 if senior}
\item{senior5}{=1 if fifth year senior}
\item{male}{=1 if male}
\item{campus}{=1 if live on campus}
\item{business}{=1 if business major}
\item{engineer}{=1 if engineering major}
\item{colGPA}{MSU GPA}
\item{hsGPA}{high school GPA}
\item{ACT}{'achievement' score}
\item{job19}{=1 if job <= 19 hours}
\item{job20}{=1 if job >= 20 hours}
\item{drive}{=1 if drive to campus}
\item{bike}{=1 if bicycle to campus}
\item{walk}{=1 if walk to campus}
\item{voluntr}{=1 if do volunteer work}
\item{PC}{=1 of pers computer at sch}
\item{greek}{=1 if fraternity or sorority}
\item{car}{=1 if own car}
\item{siblings}{=1 if have siblings}
\item{bgfriend}{=1 if boy- or girlfriend}
\item{clubs}{=1 if belong to MSU club}
\item{skipped}{avg lectures missed per week}
\item{alcohol}{avg # days per week drink alcohol}
\item{gradMI}{=1 if Michigan high school}
\item{fathcoll}{=1 if father college grad}
\item{mothcoll}{=1 if mother college grad}
}
}
\source{
Jeffrey M. Wooldrige (2006): \emph{Introductory Econometrics: A Modern Approach},
3rd ed., Thomson South-Western.
}
\usage{
data("gpa1")
}
\description{
Gollege GPA and its predictors.
}
\details{
Christopher Lemmon, a former MSU undergraduate, collected these data
from a survey he took of MSU students in Fall 1994.
}
\section{Notes}{

This is a nice example of how students can obtain an original data set
by focusing locally and carefully composing a survey.
}

\keyword{datasets}
