% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hprice1}
\alias{hprice1}
\title{hprice1}
\format{
A data frame with 88 observations.
\describe{
\item{price}{house price, $1000s}
\item{assess}{assessed value, $1000s}
\item{bdrms}{number of bedrooms}
\item{lotsize}{size of lot in square feet}
\item{sqrft}{size of house in square feet}
\item{colonial}{=1 if home is colonial style}
}
}
\source{
Wooldrige (2006): \emph{Introductory Econometrics: A Modern Approach},
3rd ed., Thomson South-Western.
}
\usage{
data("hprice1")
}
\description{
House prices and characteristics.
}
\details{
Collected from the real estate pages of the Boston Globe during 1990.
These homes sold in the Boston, MA area.
}
\section{Notes}{

Typically, it is very easy to obtain data on selling prices and
characteristics of homes, using publicly available databases.  It is
interesting to match the information on houses with other
information – such as local crime rates, quality of the local schools,
pollution levels, and so on – and estimate the effects of such variables
on housing prices.
}

\keyword{datasets}
