% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hseinv}
\alias{hseinv}
\title{hseinv}
\format{
A data frame with 42 observations.
\describe{
\item{year}{1947-1988}
\item{inv}{real housing invest., millions $}
\item{pop}{population, 1000s}
\item{price}{housing price index; 1982 = 1}
}
}
\source{
Wooldrige (2009): \emph{Introductory Econometrics: A Modern
Approach}, 4th ed. South-Western, Cengage Learning.
}
\usage{
data("hseinv")
}
\description{
Data from D. McFadden (1994), “Demographics, the Housing Market, and
the Welfare of the Elderly,” in D.A. Wise (ed.), \emph{Studies in the Economics
of Aging}. Chicago: University of Chicago Press, 225-285.
}
\keyword{datasets}
