% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{intdef}
\alias{intdef}
\title{intdef}
\format{
A data frame with 49 observations.
\describe{
\item{year}{1948-1996}
\item{i3}{3 mo. T bill rate}
\item{inf}{CPI inflation rate}
\item{rec}{federal receipts, \% GDP}
\item{out}{federal outlays, \% GDP}
\item{def}{out - rec (deficit as \% GDP)}
}
}
\source{
Wooldrige (2009): \emph{Introductory Econometrics: A Modern
Approach}, 4th ed. South-Western, Cengage Learning.
}
\usage{
data("intdef")
}
\description{
Data from \emph{Economic Report of the President}, 2004,
Tables B-64, B-73, and B-79.
}
\keyword{datasets}
