% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omit_test.R
\name{omit_test}
\alias{omit_test}
\title{Wald test for the exclusion of regressors}
\usage{
omit_test(model, frml = NULL, chisq = FALSE, .vcov = NULL, ...)
}
\arguments{
\item{model}{an \code{lm}  object.}

\item{frml}{an one-sided formula specifying the terms to be tested.
If \code{NULL} the model regressors are used.}

\item{chisq}{if \code{TRUE} compute the chi-squared statistic. Else compute
the F statistic.}

\item{.vcov}{a function computing the variance of the estimates.
If \code{NULL}, \code{vcov} is used.}

\item{...}{further parameters passed to \code{.vcov}.}
}
\value{
an object of class \code{htest} with components:
\describe{
\item{statistic}{the value of the test statistic.}
\item{p.value}{the p-value of the test.}
\item{parameter}{degrees of freedom.}
\item{method}{a character string indicating what type of test was
performed.}
\item{data.name}{a character string describing the model.}
}
}
\description{
Wald test for the exclusion of regressors
}
\examples{
data("hprice1")

mod <- lm(price ~ sqrft + lotsize + bdrms, data = hprice1)
omit_test(mod)

}
