% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rdchem}
\alias{rdchem}
\title{rdchem}
\format{
A data frame with 32 observations.
\describe{
\item{rd}{R&D spending, millions $}
\item{sales}{firm sales, millions $}
\item{profits}{profits, millions $}
}
}
\source{
Wooldrige (2009): \emph{Introductory Econometrics: A Modern
Approach}, 4th ed. South-Western, Cengage Learning.
}
\usage{
data("rdchem")
}
\description{
From \emph{Businessweek} R&D Scoreboard, October 25, 1991
}
\section{Notes}{

It would be interesting to collect more recent data and see whether
the R&D/firm size relationship has changed over time.
}

\keyword{datasets}
