% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/se.R
\name{se}
\alias{se}
\title{Standard errors of coefficient estimates}
\usage{
se(model, .vcov = NULL, ...)
}
\arguments{
\item{model}{an estimated model}

\item{.vcov}{a function to compute the covariance matrix of estimates}

\item{...}{further arguments to \code{.vcov}}
}
\value{
a named vector
}
\description{
Standard errors of coefficient estimates
}
\examples{
data("hprice1")

mod <- lm(price ~ sqrft + lotsize + bdrms, data = hprice1)
se(mod)

}
