% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{traffic2}
\alias{traffic2}
\title{traffic2}
\format{
A data frame with 108 observations.
\describe{
\item{year}{1981 to 1989}
\item{month}{1 to 12}
\item{totacc}{statewide total accidents}
\item{fatacc}{statewide fatal accidents}
\item{injacc}{statewide injury accidents}
\item{pdoacc}{property damage only accidents}
\item{ntotacc}{noninterstate total acc.}
\item{nfatacc}{noninterstate fatal acc.}
\item{ninjacc}{noninterstate injur acc.}
\item{npdoacc}{noninterstate property acc.}
\item{rtotacc}{total acc. on rural 65 mph roads}
\item{rfatacc}{fatal acc. on rural 65 mph roads}
\item{rinjacc}{injury acc. on rural 65 mph roads}
\item{rpdoacc}{property acc. on rural 65 mph roads}
\item{ushigh}{acc. on U.S. highways}
\item{cntyrds}{acc. on county roads}
\item{strtes}{acc. on state routes}
\item{unem}{state unemployment rate}
\item{spdlaw}{=1 after 65 mph in effect}
\item{beltlaw}{=1 after seatbelt law}
\item{wkends}{number of weekend days in month}
}
}
\source{
Wooldrige (2009) \emph{Introductory Econometrics: A Modern
Approach}, 4th ed. South-Western, Cengage Learning.
}
\usage{
data("traffic2")
}
\description{
Data used in: P.S. McCarthy (1994), “Relaxed Speed Limits and Highway
Safety: New Evidence from California,” \emph{Economics Letters}
\strong{46}, 173--179. Professor McCarthy kindly provided the data.
}
\section{Notes}{

Many states have changed maximum speed limits and imposed seat belt laws
over the past 25 years. Data similar to those in \code{traffic2} should be
fairly easy to obtain for a particular state. One should combine this
information with changes in a state’s blood alcohol limit and the passage
of per se and open container laws.
}

\keyword{datasets}
