% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wageprc}
\alias{wageprc}
\title{wageprc}
\format{
A data frame with 286 observations.
\describe{
\item{year}{1964 to 1987}
\item{month}{1 to 12}
\item{price}{consumer price index}
\item{wage}{nominal hourly wage}
}
}
\source{
Jeffrey M. Wooldrige (2006): \emph{Introductory Econometrics: A Modern Approach},
3rd ed., Thomson South-Western.
}
\usage{
data("wageprc")
}
\description{
Wages and prices (macro).
}
\details{
Data from \emph{Economic Report of the President}, various years.
}
\section{Notes}{

These monthly data run from January 1964 through October 1987.
The consumer price index averages to 100 in 1967. An updated
set of data can be obtained electronically from
\url{https://www.govinfo.gov/app/collection/ERP/}.
}

\keyword{datasets}
