% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef_table.R
\name{coef_table}
\alias{coef_table}
\title{Print coefficient estimates table}
\usage{
coef_table(model, vce = NULL)
}
\arguments{
\item{model}{an estimated model returned by \code{lm} or similar functions.}

\item{vce}{an object indicating how to obtain the covariance matrix.}
}
\value{
Invisibly returns the coefficient table
}
\description{
Display the output of models estimated with \link[stats:lm]{lm()} or similar
functions. The parameter \code{vce} allows to use covariance matrices consistent
to heteroskedasticity and autocorrelation (see the documentation of \link[=se]{se()}).
}
\examples{
data("hprice1")

mod <- lm(price ~ sqrft + bdrms, data = hprice1)
coef_table(mod)

# Heteroskedasticity consistent standard errors
coef_table(mod, vce = "HC")

}
