% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef_table.R
\name{coef_table}
\alias{coef_table}
\title{Print coefficient estimates table}
\usage{
coef_table(mod, .vcov = NULL, ..., digits = max(3L, getOption("digits") - 3L))
}
\arguments{
\item{mod}{an estimated model}

\item{.vcov}{a function to compute the covariance matrix of estimates}

\item{...}{further arguments to \code{.vcov}}

\item{digits}{number of digits to print}
}
\value{
invisibly returns the coefficient table
}
\description{
Print coefficient estimates table
}
\examples{
data("hprice1")

mod <- lm(price ~ sqrft + lotsize + bdrms, data = hprice1)
coef_table(mod)

}
