% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/het_test.R
\name{het_test}
\alias{het_test}
\title{LM test for heteroskedasticity}
\usage{
het_test(model, var = NULL, chisq = FALSE)
}
\arguments{
\item{model}{an \code{lm}  object.}

\item{var}{an one-sided formula for the variance function. If \code{NULL}
the model regressors are used.}

\item{chisq}{if \code{TRUE} compute the chi-squared statistic. Else compute
the F statistic.}
}
\value{
an object of class \code{htest} with components:
\describe{
\item{statistic}{the value of the test statistic.}
\item{p.value}{the p-value of the test.}
\item{parameter}{degrees of freedom.}
\item{method}{a character string indicating what type of test was
performed.}
\item{data.name}{a character string describing the model.}
}
}
\description{
LM test for heteroskedasticity
}
\examples{
data("hprice1")

mod <- lm(price ~ sqrft + lotsize + bdrms, data = hprice1)
het_test(mod)

}
