% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/se.R
\name{se}
\alias{se}
\title{Standard errors of coefficient estimates}
\usage{
se(model, vce = NULL)
}
\arguments{
\item{model}{an estimated model}

\item{vce}{a function to compute the covariance matrix of estimates}
}
\value{
a named vector
}
\description{
Standard errors of coefficient estimates
}
\examples{
data("hprice1")

mod <- lm(price ~ sqrft + lotsize + bdrms, data = hprice1)
se(mod)
# Get heteroskedasticity robust standard errors
if (require(sandwich)) {
  se(mod, vcovHC)
}

}
