% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/white_test.R
\name{white_test}
\alias{white_test}
\title{White's heteroskedasticity test}
\usage{
white_test(model, full = TRUE, chisq = FALSE)
}
\arguments{
\item{model}{an \code{lm}  object.}

\item{full}{if \code{TRUE} add the regressors, their squares and their
cross products to the auxiliary regression. If \code{FALSE} use only the
regressors and their squares.}

\item{chisq}{if \code{TRUE} compute the chi-squared statistic. Else compute
the F statistic.}
}
\value{
an object of class \code{htest} with components:
\describe{
\item{statistic}{the value of the test statistic.}
\item{p.value}{the p-value of the test.}
\item{parameter}{degrees of freedom.}
\item{method}{a character string indicating what type of test was
performed.}
\item{data.name}{a character string describing the model.}
}
}
\description{
Compute the heteroskedasticity test proposed by White, H. (1980): "A
Heteroskedasticity-Consistent Covariance Matrix Estimator and a Direct
Test for Heteroskedasticity". \emph{Econometrica}. \strong{48}(4),
817--838.
}
\examples{
data("hprice1")

mod <- lm(price ~ sqrft + lotsize + bdrms, data = hprice1)
white_test(mod)

}
